Struct casper_rust_wasm_sdk::SDK

pub struct SDK { /* private fields */ }

Implementations

impl SDK

pub async fn deploy( &self, deploy_params: DeployStrParams, session_params: SessionStrParams, payment_params: PaymentStrParams, verbosity: Option<Verbosity>, node_address: Option<String> ) -> Result<SuccessResponse<_PutDeployResult>, SdkError>

Perform a deploy operation.

Arguments

deploy_params - Deploy parameters.

session_params - Session parameters.

payment_params - Payment parameters.

verbosity - An optional verbosity level.

node_address - An optional node address.

Returns

A result containing a SuccessResponse or an SdkError.

impl SDK

pub async fn speculative_deploy( &self, deploy_params: DeployStrParams, session_params: SessionStrParams, payment_params: PaymentStrParams, maybe_block_identifier: Option<BlockIdentifierInput>, verbosity: Option<Verbosity>, node_address: Option<String> ) -> Result<SuccessResponse<_SpeculativeExecResult>, SdkError>

This function allows executing a deploy speculatively.

Arguments

deploy_params - Deployment parameters for the deploy.

session_params - Session parameters for the deploy.

payment_params - Payment parameters for the deploy.

maybe_block_identifier - Optional block identifier.

verbosity - Optional verbosity level.

node_address - Optional node address.

Returns

A Result containing either a SuccessResponse<SpeculativeExecResult> or a SdkError in case of an error.

impl SDK

pub async fn speculative_transfer( &self, amount: &str, target_account: &str, transfer_id: Option<String>, deploy_params: DeployStrParams, payment_params: PaymentStrParams, maybe_block_identifier: Option<BlockIdentifierInput>, verbosity: Option<Verbosity>, node_address: Option<String> ) -> Result<SuccessResponse<_SpeculativeExecResult>, SdkError>

Perform a speculative transfer.

Arguments

amount - The amount to transfer.

target_account - The target account.

transfer_id - An optional transfer ID (defaults to a random number).

deploy_params - The deployment parameters.

payment_params - The payment parameters.

maybe_block_identifier - An optional block identifier.

verbosity - The verbosity level for logging (optional).

node_address - The address of the node to connect to (optional).

Returns

A Result containing the result of the speculative transfer or a SdkError in case of an error.

impl SDK

pub async fn transfer( &self, amount: &str, target_account: &str, transfer_id: Option<String>, deploy_params: DeployStrParams, payment_params: PaymentStrParams, verbosity: Option<Verbosity>, node_address: Option<String> ) -> Result<SuccessResponse<_PutDeployResult>, SdkError>

Perform a transfer of funds.

Arguments

amount - The amount to transfer.

target_account - The target account.

transfer_id - An optional transfer ID (defaults to a random number).

deploy_params - The deployment parameters.

payment_params - The payment parameters.

verbosity - The verbosity level for logging (optional).

node_address - The address of the node to connect to (optional).

Returns

A Result containing the result of the transfer or a SdkError in case of an error.

impl SDK

pub async fn get_account( &self, account_identifier: Option<AccountIdentifier>, account_identifier_as_string: Option<String>, maybe_block_identifier: Option<BlockIdentifierInput>, verbosity: Option<Verbosity>, node_address: Option<String> ) -> Result<SuccessResponse<_GetAccountResult>, SdkError>

Retrieves account information based on the provided options.

Arguments

account_identifier - An optional AccountIdentifier for specifying the account identifier.

account_identifier_as_string - An optional string representing the account identifier.

maybe_block_identifier - An optional BlockIdentifierInput for specifying a block identifier.

verbosity - An optional Verbosity level for controlling the output verbosity.

node_address - An optional string specifying the node address to use for the request.

Returns

A Result containing either a SuccessResponse<_GetAccountResult> or a SdkError in case of an error.

Errors

Returns a SdkError if there is an error during the retrieval process.

impl SDK

pub async fn get_auction_info( &self, maybe_block_identifier: Option<BlockIdentifierInput>, verbosity: Option<Verbosity>, node_address: Option<String> ) -> Result<SuccessResponse<_GetAuctionInfoResult>, SdkError>

Retrieves auction information based on the provided options.

Arguments

maybe_block_identifier - An optional BlockIdentifierInput for specifying a block identifier.

verbosity - An optional Verbosity level for controlling the output verbosity.

node_address - An optional string specifying the node address to use for the request.

Returns

A Result containing either a GetAuctionInfoResult or a SdkError in case of an error.

Errors

Returns a SdkError if there is an error during the retrieval process.

impl SDK

pub async fn get_balance( &self, state_root_hash: impl ToDigest, purse_uref: GetBalanceInput, verbosity: Option<Verbosity>, node_address: Option<String> ) -> Result<SuccessResponse<_GetBalanceResult>, SdkError>

Retrieves balance information based on the provided options.

Arguments

state_root_hash - The state root hash to query for balance information.

purse_uref - The purse uref specifying the purse for which to retrieve the balance.

verbosity - An optional Verbosity level for controlling the output verbosity.

node_address - An optional string specifying the node address to use for the request.

Returns

A Result containing either a GetBalanceResult or a SdkError in case of an error.

Errors

Returns a SdkError if there is an error during the retrieval process.

impl SDK

pub async fn get_block( &self, maybe_block_identifier: Option<BlockIdentifierInput>, verbosity: Option<Verbosity>, node_address: Option<String> ) -> Result<SuccessResponse<_GetBlockResult>, SdkError>

Retrieves block information using the provided options.

Arguments

maybe_block_identifier - An optional BlockIdentifierInput specifying the block identifier.

verbosity - An optional Verbosity level for the retrieval.

node_address - An optional node address to target for retrieval.

Returns

A Result containing either a GetBlockResult or a SdkError in case of an error.

Errors

Returns a SdkError if there is an error during the retrieval process.

impl SDK

pub async fn get_block_transfers( &self, maybe_block_identifier: Option<BlockIdentifierInput>, verbosity: Option<Verbosity>, node_address: Option<String> ) -> Result<SuccessResponse<_GetBlockTransfersResult>, SdkError>

Retrieves block transfers information based on the provided options.

Arguments

maybe_block_identifier - An optional BlockIdentifierInput specifying the block identifier.

verbosity - An optional Verbosity level for controlling the output verbosity.

node_address - An optional string specifying the node address to use for the request.

Returns

A Result containing either a GetBlockTransfersResult or a SdkError in case of an error.

Errors

Returns a SdkError if there is an error during the retrieval process.

impl SDK

Implementations for the SDK struct.

pub async fn get_chainspec( &self, verbosity: Option<Verbosity>, node_address: Option<String> ) -> Result<SuccessResponse<_GetChainspecResult>, Error>

Asynchronously retrieves the chainspec.

Arguments

verbosity - An optional Verbosity parameter.

node_address - An optional node address as a string.

Returns

A Result containing either a GetChainspecResult or a SdkError in case of an error.

impl SDK

pub async fn get_deploy( &self, deploy_hash: DeployHash, finalized_approvals: Option<bool>, verbosity: Option<Verbosity>, node_address: Option<String> ) -> Result<SuccessResponse<_GetDeployResult>, Error>

Retrieves deploy information based on the provided options.

Arguments

deploy_hash - The deploy hash.

finalized_approvals - An optional boolean indicating finalized approvals.

verbosity - An optional verbosity level.

node_address - An optional node address.

Returns

A Result containing either a GetDeployResult or an error.

impl SDK

pub async fn get_dictionary_item( &self, state_root_hash: impl ToDigest, dictionary_item_input: DictionaryItemInput, verbosity: Option<Verbosity>, node_address: Option<String> ) -> Result<SuccessResponse<_GetDictionaryItemResult>, SdkError>

Retrieves dictionary item information based on the provided options.

Arguments

state_root_hash - A ToDigest implementation for specifying the state root hash.

dictionary_item - A DictionaryItemInput enum specifying the dictionary item to retrieve.

verbosity - An optional Verbosity level for controlling the output verbosity.

node_address - An optional string specifying the node address to use for the request.

Returns

A Result containing either a GetDictionaryItemResult or a SdkError in case of an error.

Errors

Returns a SdkError if there is an error during the retrieval process.

impl SDK

pub async fn get_era_info( &self, maybe_block_identifier: Option<BlockIdentifierInput>, verbosity: Option<Verbosity>, node_address: Option<String> ) -> Result<SuccessResponse<_GetEraInfoResult>, SdkError>

👎Deprecated: prefer ‘get_era_summary’ as it doesn’t require a switch block

impl SDK

pub async fn get_era_summary( &self, maybe_block_identifier: Option<BlockIdentifierInput>, verbosity: Option<Verbosity>, node_address: Option<String> ) -> Result<SuccessResponse<_GetEraSummaryResult>, SdkError>

Retrieves era summary information based on the provided options.

Arguments

maybe_block_identifier - An optional BlockIdentifierInput for specifying a block identifier.

verbosity - An optional Verbosity level for controlling the output verbosity.

node_address - An optional string specifying the node address to use for the request.

Returns

A Result containing either a GetEraSummaryResult or a SdkError in case of an error.

Errors

Returns a SdkError if there is an error during the retrieval process.

impl SDK

pub async fn get_node_status( &self, verbosity: Option<Verbosity>, node_address: Option<String> ) -> Result<SuccessResponse<_GetNodeStatusResult>, Error>

Retrieves node status information based on the provided options.

Arguments

verbosity - An optional Verbosity level for controlling the output verbosity.

node_address - An optional string specifying the node address to use for the request.

Returns

A Result containing either a GetNodeStatusResult or an Error in case of an error.

Errors

Returns an Error if there is an error during the retrieval process.

impl SDK

pub async fn get_peers( &self, verbosity: Option<Verbosity>, node_address: Option<String> ) -> Result<SuccessResponse<_GetPeersResult>, Error>

Retrieves peers.

Arguments

verbosity - Optional verbosity level.

node_address - Optional node address.

Returns

A Result containing SuccessResponse with _GetPeersResult or an Error if an error occurs.

impl SDK

pub async fn get_state_root_hash( &self, maybe_block_identifier: Option<BlockIdentifierInput>, verbosity: Option<Verbosity>, node_address: Option<String> ) -> Result<SuccessResponse<_GetStateRootHashResult>, SdkError>

Retrieves state root hash information based on the provided options.

Arguments

maybe_block_identifier - An optional BlockIdentifierInput for specifying a block identifier.

verbosity - An optional Verbosity level for controlling the output verbosity.

node_address - An optional string specifying the node address to use for the request.

Returns

A Result containing either a GetStateRootHashResult or a SdkError in case of an error.

Errors

Returns a SdkError if there is an error during the retrieval process.

impl SDK

pub async fn get_validator_changes( &self, verbosity: Option<Verbosity>, node_address: Option<String> ) -> Result<SuccessResponse<_GetValidatorChangesResult>, Error>

Retrieves validator changes based on the provided options.

Arguments

verbosity - An optional Verbosity level for controlling the output verbosity.

node_address - An optional string specifying the node address to use for the request.

Returns

A Result containing either a GetValidatorChangesResult or an Error in case of an error.

Errors

Returns an Error if there is an error during the retrieval process.

impl SDK

pub async fn list_rpcs( &self, verbosity: Option<Verbosity>, node_address: Option<String> ) -> Result<SuccessResponse<_ListRpcsResult>, Error>

Lists available RPCs based on the provided options.

Arguments

verbosity - An optional Verbosity level for controlling the output verbosity.

node_address - An optional string specifying the node address to use for the request.

Returns

A Result containing either a ListRpcsResult or an Error in case of an error.

Errors

Returns an Error if there is an error during the listing process.

impl SDK

pub async fn put_deploy( &self, deploy: Deploy, verbosity: Option<Verbosity>, node_address: Option<String> ) -> Result<SuccessResponse<_PutDeployResult>, Error>

Puts a deploy based on the provided options.

Arguments

deploy - The Deploy object to be sent.

verbosity - An optional Verbosity level for controlling the output verbosity.

node_address - An optional string specifying the node address to use for the request.

Returns

A Result containing either a PutDeployResult or an Error in case of an error.

Errors

Returns an Error if there is an error during the deploy process.

impl SDK

pub async fn query_balance( &self, maybe_global_state_identifier: Option<GlobalStateIdentifier>, purse_identifier_as_string: Option<String>, purse_identifier: Option<PurseIdentifier>, state_root_hash: Option<String>, maybe_block_id: Option<String>, verbosity: Option<Verbosity>, node_address: Option<String> ) -> Result<SuccessResponse<_QueryBalanceResult>, SdkError>

Retrieves balance information based on the provided options.

Arguments

maybe_global_state_identifier - An optional GlobalStateIdentifier for specifying global state.

purse_identifier_as_string - An optional string representing a purse identifier.

purse_identifier - An optional PurseIdentifier.

state_root_hash - An optional string representing a state root hash.

maybe_block_id - An optional string representing a block identifier.

verbosity - An optional Verbosity level for controlling the output verbosity.

node_address - An optional string specifying the node address to use for the request.

Returns

A Result containing either a SuccessResponse<_QueryBalanceResult> or a SdkError in case of an error.

Errors

Returns a SdkError if there is an error during the retrieval process.

impl SDK

pub fn query_global_state_js_alias_params( &self, options: Option<QueryGlobalStateOptions> ) -> Result<QueryGlobalStateParams, SdkError>

Builds parameters for querying global state based on the provided options.

Arguments

options - An optional QueryGlobalStateOptions struct containing retrieval options.

Returns

A Result containing either a QueryGlobalStateParams struct or a SdkError in case of an error.

pub async fn query_global_state( &self, query_params: QueryGlobalStateParams ) -> Result<SuccessResponse<_QueryGlobalStateResult>, SdkError>

Retrieves global state information based on the provided parameters.

Arguments

query_params - A QueryGlobalStateParams struct containing query parameters.

Returns

A Result containing either a SuccessResponse<_QueryGlobalStateResult> or a SdkError in case of an error.

impl SDK

pub async fn speculative_exec( &self, deploy: Deploy, maybe_block_identifier: Option<BlockIdentifierInput>, verbosity: Option<Verbosity>, node_address: Option<String> ) -> Result<SuccessResponse<_SpeculativeExecResult>, SdkError>

Perform speculative execution.

Arguments

deploy - The deploy to execute.

maybe_block_identifier - The block identifier.

verbosity - The verbosity level for logging.

node_address - The address of the node to connect to.

Returns

A Result containing the result of the speculative execution or a SdkError in case of an error.

impl SDK

pub fn make_deploy( &self, deploy_params: DeployStrParams, session_params: SessionStrParams, payment_params: PaymentStrParams ) -> Result<Deploy, SdkError>

Creates a deploy using the provided parameters.

Arguments

deploy_params - The deploy parameters.

session_params - The session parameters.

payment_params - The payment parameters.

Returns

A Result containing the created Deploy or a SdkError in case of an error.

impl SDK

pub fn make_transfer( &self, amount: &str, target_account: &str, transfer_id: Option<String>, deploy_params: DeployStrParams, payment_params: PaymentStrParams ) -> Result<Deploy, SdkError>

Creates a transfer deploy with the provided parameters.

Arguments

amount - The transfer amount.

target_account - The target account.

transfer_id - Optional transfer identifier.

deploy_params - The deploy parameters.

payment_params - The payment parameters.

Returns

A Result containing the created Deploy or a SdkError in case of an error.

impl SDK

pub fn sign_deploy(&self, deploy: Deploy, secret_key: &str) -> Deploy

Signs a deploy using the provided secret key.

Arguments

deploy - The deploy to sign.

secret_key - The secret key for signing.

Returns

The signed Deploy.

impl SDK

pub fn watch_deploy( &self, events_url: &str, timeout_duration: Option<u64> ) -> DeployWatcher

Creates a new DeployWatcher instance to watch deploys.

Arguments

events_url - The URL to monitor for deploy events.

timeout_duration - An optional timeout duration in seconds.

Returns

A DeployWatcher instance.

pub fn watch_deploy_js_alias( &self, events_url: &str, timeout_duration: Option<u32> ) -> DeployWatcher

Creates a new DeployWatcher instance to watch deploys (JavaScript-friendly).

Arguments

events_url - The URL to monitor for deploy events.

timeout_duration - An optional timeout duration in seconds.

Returns

A DeployWatcher instance.

pub async fn wait_deploy( &self, events_url: &str, deploy_hash: &str, timeout_duration: Option<u64> ) -> Result<EventParseResult, String>

Waits for a deploy event to be processed asynchronously.

Arguments

events_url - The URL to monitor for deploy events.

deploy_hash - The deploy hash to wait for.

timeout_duration - An optional timeout duration in milliseconds.

Returns

A Result containing either the processed EventParseResult or an error message.

impl SDK

A set of functions for working with smart contract entry points. Alias of sdk.deploy

pub async fn call_entrypoint( &self, deploy_params: DeployStrParams, session_params: SessionStrParams, payment_params: PaymentStrParams, node_address: Option<String> ) -> Result<SuccessResponse<_PutDeployResult>, SdkError>

Calls a smart contract entry point with the specified parameters and returns the result.

Arguments

deploy_params - The deploy parameters.

session_params - The session parameters.

payment_params - The payment parameters.

node_address - An optional node address to send the request to.

Returns

A Result containing either a PutDeployResult or a SdkError in case of an error.

Errors

Returns a SdkError if there is an error during the call.

impl SDK

A set of functions for installing smart contracts on the blockchain. Alias of sdk.deploy

pub async fn install( &self, deploy_params: DeployStrParams, session_params: SessionStrParams, payment_params: PaymentStrParams, node_address: Option<String> ) -> Result<SuccessResponse<_PutDeployResult>, SdkError>

Installs a smart contract with the specified parameters and returns the result.

Arguments

deploy_params - The deploy parameters.

session_params - The session parameters.

payment_params - The payment parameters.

node_address - An optional node address to send the request to.

Returns

A Result containing either a PutDeployResult or a SdkError in case of an error.

Errors

Returns a SdkError if there is an error during the installation.

impl SDK

Alias of sdk.get_dictionary_item

pub async fn query_contract_dict( &self, state_root_hash: impl ToDigest, dictionary_item: DictionaryItemInput, verbosity: Option<Verbosity>, node_address: Option<String> ) -> Result<SuccessResponse<_GetDictionaryItemResult>, SdkError>

Query a contract dictionary item.

Arguments

state_root_hash - State root hash.

dictionary_item - Dictionary item input.

verbosity - Optional verbosity level.

node_address - Optional node address.

Returns

A Result containing either a SuccessResponse<GetDictionaryItemResult> or a SdkError in case of an error.

impl SDK

Alias of sdk.query_global_state

pub async fn query_contract_key( &self, query_params: QueryGlobalStateParams ) -> Result<SuccessResponse<_QueryGlobalStateResult>, SdkError>

Query a contract key.

Arguments

query_params - Query global state parameters.

Returns

A Result containing either a SuccessResponse<QueryGlobalStateResult> or a SdkError in case of an error.

impl SDK

pub fn new(node_address: Option<String>, verbosity: Option<Verbosity>) -> Self

pub fn get_node_address(&self, node_address: Option<String>) -> String

pub fn set_node_address( &mut self, node_address: Option<String> ) -> Result<(), String>

pub fn get_verbosity(&self, verbosity: Option<Verbosity>) -> Verbosity

pub fn set_verbosity( &mut self, verbosity: Option<Verbosity> ) -> Result<(), String>